Adding delete confirmation alert on history schedule

+++ CleanPlot/HistoryView.swift
@@ -12,29 +12,49 @@ import SwiftData
 struct HistoryView: View {
     @Query var schedules: [Schedule]
     @Environment(\.modelContext) var modelContext
+    @State var showDeleteAlert = false
+    @State var indexSetToDelete: IndexSet? = nil
 
     var body: some View {
         NavigationStack {
-            VStack(alignment:.leading) {
-                List {
-                    ForEach(schedules) {
-                        schedule in
-                        NavigationLink {
-                            DetailHistory(schedule: schedule)
-                        } label: {
-                            Text(schedule.scheduleId)
+            VStack {
+                if schedules.isEmpty {
+                    Text("Belum ada riwayat jadwal").foregroundStyle(.secondary)
+                }
+                else {
+                    List {
+                        ForEach(schedules) {
+                            schedule in
+                            NavigationLink {
+                                DetailHistory(schedule: schedule)
+                            } label: {
+                                Text(schedule.scheduleId)
+                            }
+                        }.onDelete { offsets in
+                            indexSetToDelete = offsets
+                            showDeleteAlert = true
                         }
-                    }.onDelete { offsets in
+                        
+                    }
+                }
+            }
+            .navigationTitle("Riwayat")
+            .alert("Hapus Data", isPresented: $showDeleteAlert) {
+                Button("Hapus", role: .destructive) {
+                    if let offsets = indexSetToDelete {
                         for offset in offsets {
                             let schedule = schedules[offset]
-                            
                             modelContext.delete(schedule)
                         }
                     }
-                    
+                    indexSetToDelete = nil
                 }
-                
-            }.navigationTitle("Riwayat")
+                Button("Batal", role: .cancel) {
+                    indexSetToDelete = nil
+                }
+            } message: {
+                Text("Apakah Anda yakin ingin menghapus data ini?")
+            }
         }
     }
 }
